Please don't sue me Discord
We've all had our up's and down's this year. Let's just go flat in the middle on this one and pretend it doesn't violate the Terms in one way or another. Reach out to me on Discord @TallerThanShort#0775 (while my nitro lasts) or @TallerThanShort#8186 (my alt that's never had nitro)